moved remove_chartjunk into utils
