+++ prettyplotlib/utils.py
@@ -1,0 +1,63 @@
+__author__ = 'olga'
+
+
+def remove_chartjunk(ax, spines, grid=None, ticklabels=None, show_ticks=False):
+    '''
+    Removes "chartjunk", such as extra lines of axes and tick marks.
+
+    If grid="y" or "x", will add a white grid at the "y" or "x" axes,
+    respectively
+
+    If ticklabels="y" or "x", or ['x', 'y'] will remove ticklabels from that
+    axis
+    '''
+    all_spines = ['top', 'bottom', 'right', 'left']
+    for spine in spines:
+        ax.spines[spine].set_visible(False)
+
+    # For the remaining spines, make their line thinner and a slightly
+    # off-black dark grey
+    for spine in all_spines:
+        if spine not in spines:
+            ax.spines[spine].set_linewidth(0.5)
+            # ax.spines[spine].set_color(almost_black)
+            #            ax.spines[spine].set_tick_params(color=almost_black)
+            # Check that the axes are not log-scale. If they are, leave the ticks
+            # because otherwise people assume a linear scale.
+    x_pos = set(['top', 'bottom'])
+    y_pos = set(['left', 'right'])
+    xy_pos = [x_pos, y_pos]
+    xy_ax_names = ['xaxis', 'yaxis']
+
+    for ax_name, pos in zip(xy_ax_names, xy_pos):
+        axis = ax.__dict__[ax_name]
+        # axis.set_tick_params(color=almost_black)
+        #print 'axis.get_scale()', axis.get_scale()
+        if show_ticks or axis.get_scale() == 'log':
+            # if this spine is not in the list of spines to remove
+            for p in pos.difference(spines):
+                #print 'p', p
+                axis.set_tick_params(direction='out')
+                axis.set_ticks_position(p)
+                #                axis.set_tick_params(which='both', p)
+        else:
+            axis.set_ticks_position('none')
+
+    if grid is not None:
+        for g in grid:
+            assert g in ('x', 'y')
+            ax.grid(axis=grid, color='white', linestyle='-', linewidth=0.5)
+
+    if ticklabels is not None:
+        if type(ticklabels) is str:
+            assert ticklabels in set(('x', 'y'))
+            if ticklabels == 'x':
+                ax.set_xticklabels([])
+            if ticklabels == 'y':
+                ax.set_yticklabels([])
+        else:
+            assert set(ticklabels) | set(('x', 'y')) > 0
+            if 'x' in ticklabels:
+                ax.set_xticklabels([])
+            elif 'y' in ticklabels:
+                ax.set_yticklabels([])
